Fix missing include and type mismatch in LaunchPosition tests

+++ src/cascadia/LocalTests_TerminalApp/SettingsTests.cpp
@@ -6,6 +6,7 @@
 #include "../TerminalApp/TerminalPage.h"
 #include "../UnitTests_SettingsModel/TestUtils.h"
 #include "../TerminalSettingsAppAdapterLib/TerminalSettings.h"
+#include "../TerminalSettingsModel/ModelSerializationHelpers.h"
 
 using namespace Microsoft::Console;
 using namespace WEX::Logging;
@@ -1627,14 +1628,14 @@ namespace TerminalAppLocalTests
             const auto pos = LaunchPositionFromString("50");
             VERIFY_IS_TRUE(pos.X.has_value());
             VERIFY_IS_TRUE(pos.Y.has_value());
-            VERIFY_ARE_EQUAL(50, pos.X.Value());
-            VERIFY_ARE_EQUAL(50, pos.Y.Value());
+            VERIFY_ARE_EQUAL(50, static_cast<int32_t>pos.X.Value());
+            VERIFY_ARE_EQUAL(50, static_cast<int32_t>pos.Y.Value());
         }
 
         {
             const auto pos = LaunchPositionFromString("100,");
             VERIFY_IS_TRUE(pos.X.has_value());
-            VERIFY_ARE_EQUAL(100, pos.X.Value());
+            VERIFY_ARE_EQUAL(100, static_cast<int32_t>pos.X.Value());
             VERIFY_IS_FALSE(pos.Y.has_value());
         }
 
@@ -1642,22 +1643,22 @@ namespace TerminalAppLocalTests
             const auto pos = LaunchPositionFromString(",100");
             VERIFY_IS_FALSE(pos.X.has_value());
             VERIFY_IS_TRUE(pos.Y.has_value());
-            VERIFY_ARE_EQUAL(100, pos.Y.Value());
+            VERIFY_ARE_EQUAL(100, static_cast<int32_t>pos.Y.Value());
         }
 
         {
             const auto pos = LaunchPositionFromString("50,50");
             VERIFY_IS_TRUE(pos.X.has_value());
-            VERIFY_ARE_EQUAL(50, pos.X.Value());
+            VERIFY_ARE_EQUAL(50, static_cast<int32_t>pos.X.Value());
             VERIFY_IS_TRUE(pos.Y.has_value());
-            VERIFY_ARE_EQUAL(50, pos.Y.Value());
+            VERIFY_ARE_EQUAL(50, static_cast<int32_t>pos.Y.Value());
         }
 
         {
             const auto pos = LaunchPositionFromString("abc,100");
             VERIFY_IS_FALSE(pos.X.has_value());
             VERIFY_IS_TRUE(pos.Y.has_value());
-            VERIFY_ARE_EQUAL(100, pos.Y.Value());
+            VERIFY_ARE_EQUAL(100, static_cast<int32_t>pos.Y.Value());
         }
 
         {
